requirements.txt
python-dotenv==1.0.1
numpy==2.2.3
pandas==2.2.3
plotly-express==0.4.1
scipy==1.15.2
streamlit==1.42.0
langchain==0.3.19
langchain-groq==0.2.4
streamlit-mathlive==0.0.2
sympy==1.13.3
latex2sympy2
antlr4-python3-runtime==4.11
